fix(core): keep authority summary mint closure-private
