+++ packages/core/test/system-record-package-export-v1.mjs
@@ -293,4 +293,13 @@ for (const internal of [
   );
 }
 
+const authoritySummaryInternal = await import(
+  '../dist/system-record-authority-summary-v1-internal.js'
+);
+assert.equal(
+  'mintAgentProfileVerifiedAuthoritySummaryV1' in authoritySummaryInternal,
+  false,
+  'deep-imported callers must not be able to mint verified authority summaries',
+);
+
 console.log(`system-record-v1 package export ok (${expectedRuntimeExports.length} exact symbols)`);
